禁止个人信息编辑空格等非法字符

+++ app/src/main/java/com/thfw/mobileheart/activity/settings/InfoActivity.java
@@ -290,6 +290,9 @@ public class InfoActivity extends BaseActivity<UserInfoPresenter> implements Use
     }
 
     private boolean isTextEmpty(TextView textView) {
+        if (textView == null) {
+            return true;
+        }
         return TextUtils.isEmpty(textView.getText().toString());
     }
 
@@ -835,10 +838,12 @@ public class InfoActivity extends BaseActivity<UserInfoPresenter> implements Use
                         if (EmptyUtil.isEmpty(InfoActivity.this)) {
                             return;
                         }
-                        // 删除源文件
-                        File file = new File(avatarUrl);
-                        if (file.exists()) {
-                            file.delete();
+                        if (!TextUtils.isEmpty(avatarUrl)) {
+                            // 删除源文件
+                            File file = new File(avatarUrl);
+                            if (file.exists()) {
+                                file.delete();
+                            }
                         }
                         LoadingDialog.hide();
                         HttpResult<HeadModel> result = GsonUtil.fromJson(json, type);
