+++ app_robot/src/main/java/com/thfw/robotheart/activitys/me/InfoActivity.java
@@ -288,6 +288,9 @@ public class InfoActivity extends RobotBaseActivity<UserInfoPresenter> implement
     }
 
     private boolean isTextEmpty(TextView textView) {
+        if (textView == null) {
+            return true;
+        }
         return TextUtils.isEmpty(textView.getText().toString());
     }
 
@@ -833,10 +836,12 @@ public class InfoActivity extends RobotBaseActivity<UserInfoPresenter> implement
                         if (EmptyUtil.isEmpty(InfoActivity.this)) {
                             return;
                         }
-                        // 删除源文件
-                        File file = new File(avatarUrl);
-                        if (file.exists()) {
-                            file.delete();
+                        if (!TextUtils.isEmpty(avatarUrl)) {
+                            // 删除源文件
+                            File file = new File(avatarUrl);
+                            if (file.exists()) {
+                                file.delete();
+                            }
                         }
                         LoadingDialog.hide();
                         HttpResult<HeadModel> result = GsonUtil.fromJson(json, type);
